requirements.txt
langchain-google-genai
langchain-community
python-dotenv
streamlit
